+++ firebase_sub/.gitignore
@@ -2,3 +2,4 @@
 .vscode
 *_venv
 token.json
+cred.json
